requirements.txt
flask
flask_socketio
gunicorn
python-dotenv
pandas
requests
